requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
passlib==1.7.4
python-dotenv==0.21.1
requests==2.28.2
gunicorn==19.0.0
psycopg2-binary==2.9.5
Jinja2==3.1.2
SQLAlchemy==2.0.7
google-search-results==2.4.2